Translate 25 more svg/animation tests to WPT
This is the second commit in the series of updating all the old svg
animation tests.

The usage of testharness has replaced the older SVGAnimationTest.js
for all where it's suitable. No functionality should have changed
and the tests should cover almost the same.

In all of the animations where there is a sampling at T=0, where
it was assumed that no animations had started. Which didn't work
flawlessly when moved to the new system, it has thus been removed.

Bug: 985335
Change-Id: If5d6d879be3ccd3b14467eebaa4c9a2c03128f64